DockerFR/requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
insightface
opencv-python
numpy
onnxruntime
scikit-learn
Pillow
